chore(explorer): suggestions
