+++ apps/explorer/src/lib/constants/index.ts
@@ -4,3 +4,4 @@
 export * from './footer.constants';
 export * from './validator.constants';
 export * from './pageSize.constants';
+export * from './policy.constants';
